Move MirrorEmbryo plugin into spots package
