+++ src/test/java/org/mastodon/mamut/tomancak/spots/MirrorEmbryoTest.java
@@ -1,4 +1,4 @@
-package org.mastodon.mamut.tomancak;
+package org.mastodon.mamut.tomancak.spots;
 
 import static org.junit.Assert.assertArrayEquals;
